change nonce to uint32

+++ blockchain/header.go
@@ -9,14 +9,14 @@ import (
 )
 
 type Header struct {
-	Nonce uint64
+	Nonce uint32
 	Prev  string
 	Root  string
 	Hash  string
 }
 
-func NewHeader(nonce uint64, prev, root string) *Header {
-	h := hash.BTCHash(prev + root + strconv.FormatUint(nonce, 10))
+func NewHeader(nonce uint32, prev, root string) *Header {
+	h := hash.BTCHash(prev + root + strconv.FormatUint(uint64(nonce), 10))
 
 	return &Header{nonce, prev, root, h}
 }
